Data.Store.Internal

Description

Internal API for the store package. The functions here which are not re-exported by Data.Store are less likely to have stable APIs.

This module also defines most of the included Store instances, for types from the base package and other commonly used packages (bytestring, containers, text, time, etc).

Encoding and decoding strict ByteStrings.

encode :: Store a => a -> ByteString

Serializes a value to a ByteString. In order to do this, it first allocates a ByteString of the correct size (based on size), and then uses poke to fill it.

Safety of this function depends on correctness of the Store instance. If size returns a. The good news is that this isn't an issue if you use well-tested manual instances (such as those from this package) combined with auomatic definition of instances.

decode :: Store a => ByteString -> Either PeekException a

Decodes a value from a ByteString. Returns an exception if there's an error while decoding, or if decoding undershoots / overshoots the end of the buffer.

decodeWith :: Peek a -> ByteString -> Either PeekException a

Decodes a value from a ByteString, potentially throwing exceptions, and taking a Peek to run. It is an exception to not consume all input.

decodeEx :: Store a => ByteString -> a

Decodes a value from a ByteString, potentially throwing exceptions. It is an exception to not consume all input.

decodeExWith :: Peek a -> ByteString -> a

Decodes a value from a ByteString, potentially throwing exceptions, and taking a Peek to run. It is an exception to not consume all input.

decodeExPortionWith :: Peek a -> ByteString -> (Offset, a)

Similar to decodeExWith, but it allows there to be more of the buffer remaining. The Offset of the buffer contents immediately after the decoded value is returned.

decodeIO :: Store a => ByteString -> IO a

Decodes a value from a ByteString, potentially throwing exceptions. It is an exception to not consume all input.

decodeIOWith :: Peek a -> ByteString -> IO a

Decodes a value from a ByteString, potentially throwing exceptions, and taking a Peek to run. It is an exception to not consume all input.

decodeIOPortionWith :: Peek a -> ByteString -> IO (Offset, a)

Similar to decodeExPortionWith, but runs in the IO monad.

Store class and related types.

class Store a where

The Store typeclass provides efficient serialization and deserialization to raw pointer addresses.

The peek and poke methods should be defined such that decodeEx (encode x) == x .

Minimal complete definition

Nothing

Methods

size :: Size a

Yields the Size of the buffer, in bytes, required to store the encoded representation of the type.

Note that the correctness of this function is crucial for the safety of poke, as it does not do any bounds checking. It is the responsibility of the invoker of poke (encode and similar functions) to ensure that there's enough space in the output buffer. If poke writes beyond, then arbitrary memory can be overwritten, causing undefined behavior and segmentation faults.

default size :: (Generic a, GStoreSize (Rep a)) => Size a

poke :: a -> Poke ()

Serializes a value to bytes. It is the responsibility of the caller to ensure that at least the number of bytes required by size are available. These details are handled by encode and similar utilities.

default poke :: (Generic a, GStorePoke (Rep a)) => a -> Poke ()

peek :: Peek a

Serialized a value from bytes, throwing exceptions if it encounters invalid data or runs out of input bytes.

default peek :: (Generic a, GStorePeek (Rep a)) => Peek a

data Poke a

Poke actions are useful for building sequential serializers.

They are actions which write values to bytes into memory specified by a Ptr base. The Applicative and Monad instances make it easy to write serializations, by keeping track of the Offset of the current byte. They allow you to chain Poke action such that subsequent Pokes write into subsequent portions of the output.

data Peek a

Peek actions are useful for building sequential deserializers.

They are actions which read from memory and construct values from it. The Applicative and Monad instances make it easy to chain these together to get more complicated deserializers. This machinery keeps track of the current Ptr and end-of-buffer Ptr.

runPeek :: Peek a -> PeekState -> Ptr Word8 -> IO (PeekResult a)

Run the Peek action, with a Ptr to the end of the buffer where data is poked, and a Ptr to the current position. The result is the Ptr, along with a return value.

May throw a PeekException if the memory contains invalid values.

Exceptions thrown by Poke

data PokeException

Exception thrown while running poke. Note that other types of exceptions could also be thrown. Invocations of fail in the Poke monad causes this exception to be thrown.

PokeExceptions are not expected to occur in ordinary circumstances, and usually indicate a programming error.

Constructors

PokeException

Fields

pokeExByteIndex :: Offset

pokeExMessage :: Text

pokeException :: Text -> Poke a

Throws a PokeException. These should be avoided when possible, they usually indicate a programming error.

Exceptions thrown by Peek

data PeekException

Exception thrown while running peek. Note that other types of exceptions can also be thrown. Invocations of fail in the Poke monad causes this exception to be thrown.

PeekException is thrown when the data being decoded is invalid.

Constructors

PeekException

Fields

peekExBytesFromEnd :: Offset

peekExMessage :: Text

peekException :: Text -> Peek a

Throws a PeekException.

tooManyBytes :: Int -> Int -> String -> IO void

Throws a PeekException about an attempt to read too many bytes.

Size type

data Size a

Info about a type's serialized length. Either the length is known independently of the value, or the length depends on the value.

Constructors

VarSize (a -> Int)

ConstSize !Int

getSize :: Store a => a -> Int

Get the number of bytes needed to store the given value. See size.

getSizeWith :: Size a -> a -> Int

Given a Size value and a value of the type a, returns its Int size.

combineSize :: forall a b c. (Store a, Store b) => (c -> a) -> (c -> b) -> Size c

Create an aggregate Size by providing functions to split the input into two pieces.

If both of the types are ConstSize, the result is ConstSize and the functions will not be used.

combineSizeWith :: forall a b c. (c -> a) -> (c -> b) -> Size a -> Size b -> Size c

Create an aggregate Size by providing functions to split the input into two pieces, as well as Size values to use to measure the results.

If both of the input Size values are ConstSize, the result is ConstSize and the functions will not be used.

addSize :: Int -> Size a -> Size a

Adds a constant amount to a Size value.

Store instances in terms of IsSequence

sizeSequence :: forall t. (IsSequence t, Store (Element t)) => Size t

Implement size for an IsSequence of Store instances.

Note that many monomorphic containers have more efficient implementations (for example, via memcpy).

pokeSequence :: (IsSequence t, Store (Element t)) => t -> Poke ()

Implement poke for an IsSequence of Store instances.

Note that many monomorphic containers have more efficient implementations (for example, via memcpy).

peekSequence :: (IsSequence t, Store (Element t), Index t ~ Int) => Peek t

Implement peek for an IsSequence of Store instances.

Note that many monomorphic containers have more efficient implementations (for example, via memcpy).

Store instances in terms of IsSet

sizeSet :: forall t. (IsSet t, Store (Element t)) => Size t

Implement size for an IsSet of Store instances.

pokeSet :: (IsSet t, Store (Element t)) => t -> Poke ()

Implement poke for an IsSequence of Store instances.

peekSet :: (IsSet t, Store (Element t)) => Peek t

Implement peek for an IsSequence of Store instances.

Store instances in terms of IsMap

sizeMap :: forall t. (Store (ContainerKey t), Store (MapValue t), IsMap t) => Size t

Implement size for an IsMap of where both ContainerKey and MapValue are Store instances.

pokeMap :: (Store (ContainerKey t), Store (MapValue t), IsMap t) => t -> Poke ()

Implement poke for an IsMap of where both ContainerKey and MapValue are Store instances.

peekMap :: (Store (ContainerKey t), Store (MapValue t), IsMap t) => Peek t

Implement peek for an IsMap of where both ContainerKey and MapValue are Store instances.

Utilities for ordered maps

sizeOrdMap :: forall t. (Store (ContainerKey t), Store (MapValue t), IsMap t) => Size t

Like sizeMap but should only be used for ordered containers where mapToList returns an ascending list.

pokeOrdMap :: (Store (ContainerKey t), Store (MapValue t), IsMap t) => t -> Poke ()

Like pokeMap but should only be used for ordered containers where mapToList returns an ascending list.

peekOrdMapWith

Arguments

:: (Store (ContainerKey t), Store (MapValue t))

=> ([(ContainerKey t, MapValue t)] -> t)
A function to construct the map from an ascending list such as fromDistinctAscList.

-> Peek t

Decode the results of pokeOrdMap using a given function to construct the map.

Store instances in terms of IArray

sizeArray :: (Ix i, IArray a e, Store i, Store e) => Size (a i e)

pokeArray :: (Ix i, IArray a e, Store i, Store e) => a i e -> Poke ()

peekArray :: (Ix i, IArray a e, Store i, Store e) => Peek (a i e)

Store instances in terms of Generic

class GStoreSize f

Minimal complete definition

gsize

genericSize :: (Generic a, GStoreSize (Rep a)) => Size a

class GStorePoke f

Minimal complete definition

gpoke

genericPoke :: (Generic a, GStorePoke (Rep a)) => a -> Poke ()

class GStorePeek f

Minimal complete definition

gpeek

genericPeek :: (Generic a, GStorePeek (Rep a)) => Peek a

Peek utilities

skip :: Int -> Peek ()

Skip n bytes forward.

isolate :: Int -> Peek a -> Peek a

Isolate the input to n bytes, skipping n bytes forward. Fails if m advances the offset beyond the isolated region.

peekMagic :: (Eq a, Show a, Store a) => String -> a -> Peek ()

Ensure the presence of a given magic value.

Throws a PeekException if the value isn't present.

Static Size type

class KnownNat n => IsStaticSize n a where

Methods

toStaticSize :: a -> Maybe (StaticSize n a)

newtype StaticSize (n :: Nat) a

Constructors

StaticSize

Fields

unStaticSize :: a

toStaticSizeEx :: IsStaticSize n a => a -> StaticSize n a

liftStaticSize :: forall n a. (KnownNat n, Lift a) => TypeQ -> StaticSize n a -> ExpQ

staticByteStringExp :: ByteString -> ExpQ

Orphan instances

Store Bool

Instance details

Methods

size :: Size Bool

poke :: Bool -> Poke ()

peek :: Peek Bool

Store Char

Instance details

Methods

size :: Size Char

poke :: Char -> Poke ()

peek :: Peek Char

Store Double

Instance details

Methods

size :: Size Double

poke :: Double -> Poke ()

peek :: Peek Double

Store Float

Instance details

Methods

size :: Size Float

poke :: Float -> Poke ()

peek :: Peek Float

Store Int

Instance details

Methods

size :: Size Int

poke :: Int -> Poke ()

peek :: Peek Int

Store Int8

Instance details

Methods

size :: Size Int8

poke :: Int8 -> Poke ()

peek :: Peek Int8

Store Int16

Instance details

Methods

size :: Size Int16

poke :: Int16 -> Poke ()

peek :: Peek Int16

Store Int32

Instance details

Methods

size :: Size Int32

poke :: Int32 -> Poke ()

peek :: Peek Int32

Store Int64

Instance details

Methods

size :: Size Int64

poke :: Int64 -> Poke ()

peek :: Peek Int64

Store Integer

Instance details

Methods

size :: Size Integer

poke :: Integer -> Poke ()

peek :: Peek Integer

Store Natural

Instance details

Methods

size :: Size Natural

poke :: Natural -> Poke ()

peek :: Peek Natural

Store Word

Instance details

Methods

size :: Size Word

poke :: Word -> Poke ()

peek :: Peek Word

Store Word8

Instance details

Methods

size :: Size Word8

poke :: Word8 -> Poke ()

peek :: Peek Word8

Store Word16

Instance details

Methods

size :: Size Word16

poke :: Word16 -> Poke ()

peek :: Peek Word16

Store Word32

Instance details

Methods

size :: Size Word32

poke :: Word32 -> Poke ()

peek :: Peek Word32

Store Word64

Instance details

Methods

size :: Size Word64

poke :: Word64 -> Poke ()

peek :: Peek Word64

Store Exp

Instance details

Methods

size :: Size Exp

poke :: Exp -> Poke ()

peek :: Peek Exp

Store Match

Instance details

Methods

size :: Size Match

poke :: Match -> Poke ()

peek :: Peek Match

Store Clause

Instance details

Methods

size :: Size Clause

poke :: Clause -> Poke ()

peek :: Peek Clause

Store Pat

Instance details

Methods

size :: Size Pat

poke :: Pat -> Poke ()

peek :: Peek Pat

Store Type

Instance details

Methods

size :: Size Type

poke :: Type -> Poke ()

peek :: Peek Type

Store Dec

Instance details

Methods

size :: Size Dec

poke :: Dec -> Poke ()

peek :: Peek Dec

Store Name

Instance details

Methods

size :: Size Name

poke :: Name -> Poke ()

peek :: Peek Name

Store FunDep

Instance details

Methods

size :: Size FunDep

poke :: FunDep -> Poke ()

peek :: Peek FunDep

Store InjectivityAnn

Instance details

Methods

size :: Size InjectivityAnn

poke :: InjectivityAnn -> Poke ()

peek :: Peek InjectivityAnn

Store Overlap

Instance details

Methods

size :: Size Overlap

poke :: Overlap -> Poke ()

peek :: Peek Overlap

Store ()

Instance details

Methods

size :: Size ()

poke :: () -> Poke ()

peek :: Peek ()

Store Void

Instance details

Methods

size :: Size Void

poke :: Void -> Poke ()

peek :: Peek Void

Store CDev

Instance details

Methods

size :: Size CDev

poke :: CDev -> Poke ()

peek :: Peek CDev

Store CIno

Instance details

Methods

size :: Size CIno

poke :: CIno -> Poke ()

peek :: Peek CIno

Store CMode

Instance details

Methods

size :: Size CMode

poke :: CMode -> Poke ()

peek :: Peek CMode

Store COff

Instance details

Methods

size :: Size COff

poke :: COff -> Poke ()

peek :: Peek COff

Store CPid

Instance details

Methods

size :: Size CPid

poke :: CPid -> Poke ()

peek :: Peek CPid

Store CSsize

Instance details

Methods

size :: Size CSsize

poke :: CSsize -> Poke ()

peek :: Peek CSsize

Store CGid

Instance details

Methods

size :: Size CGid

poke :: CGid -> Poke ()

peek :: Peek CGid

Store CNlink

Instance details

Methods

size :: Size CNlink

poke :: CNlink -> Poke ()

peek :: Peek CNlink

Store CUid

Instance details

Methods

size :: Size CUid

poke :: CUid -> Poke ()

peek :: Peek CUid

Store CCc

Instance details

Methods

size :: Size CCc

poke :: CCc -> Poke ()

peek :: Peek CCc

Store CSpeed

Instance details

Methods

size :: Size CSpeed

poke :: CSpeed -> Poke ()

peek :: Peek CSpeed

Store CTcflag

Instance details

Methods

size :: Size CTcflag

poke :: CTcflag -> Poke ()

peek :: Peek CTcflag

Store CRLim

Instance details

Methods

size :: Size CRLim

poke :: CRLim -> Poke ()

peek :: Peek CRLim

Store CBlkSize

Instance details

Methods

size :: Size CBlkSize

poke :: CBlkSize -> Poke ()

peek :: Peek CBlkSize

Store CBlkCnt

Instance details

Methods

size :: Size CBlkCnt

poke :: CBlkCnt -> Poke ()

peek :: Peek CBlkCnt

Store CClockId

Instance details

Methods

size :: Size CClockId

poke :: CClockId -> Poke ()

peek :: Peek CClockId

Store CFsBlkCnt

Instance details

Methods

size :: Size CFsBlkCnt

poke :: CFsBlkCnt -> Poke ()

peek :: Peek CFsBlkCnt

Store CFsFilCnt

Instance details

Methods

size :: Size CFsFilCnt

poke :: CFsFilCnt -> Poke ()

peek :: Peek CFsFilCnt

Store CId

Instance details

Methods

size :: Size CId

poke :: CId -> Poke ()

peek :: Peek CId

Store CKey

Instance details

Methods

size :: Size CKey

poke :: CKey -> Poke ()

peek :: Peek CKey

Store CTimer

Instance details

Methods

size :: Size CTimer

poke :: CTimer -> Poke ()

peek :: Peek CTimer

Store CSocklen

Instance details

Methods

size :: Size CSocklen

poke :: CSocklen -> Poke ()

peek :: Peek CSocklen

Store CNfds

Instance details

Methods

size :: Size CNfds

poke :: CNfds -> Poke ()

peek :: Peek CNfds

Store Fd

Instance details

Methods

size :: Size Fd

poke :: Fd -> Poke ()

peek :: Peek Fd

Store All

Instance details

Methods

size :: Size All

poke :: All -> Poke ()

peek :: Peek All

Store Any

Instance details

Methods

size :: Size Any

poke :: Any -> Poke ()

peek :: Peek Any

Store CChar

Instance details

Methods

size :: Size CChar

poke :: CChar -> Poke ()

peek :: Peek CChar

Store CSChar

Instance details

Methods

size :: Size CSChar

poke :: CSChar -> Poke ()

peek :: Peek CSChar

Store CUChar

Instance details

Methods

size :: Size CUChar

poke :: CUChar -> Poke ()

peek :: Peek CUChar

Store CShort

Instance details

Methods

size :: Size CShort

poke :: CShort -> Poke ()

peek :: Peek CShort

Store CUShort

Instance details

Methods

size :: Size CUShort

poke :: CUShort -> Poke ()

peek :: Peek CUShort

Store CInt

Instance details

Methods

size :: Size CInt

poke :: CInt -> Poke ()

peek :: Peek CInt

Store CUInt

Instance details

Methods

size :: Size CUInt

poke :: CUInt -> Poke ()

peek :: Peek CUInt

Store CLong

Instance details

Methods

size :: Size CLong

poke :: CLong -> Poke ()

peek :: Peek CLong

Store CULong

Instance details

Methods

size :: Size CULong

poke :: CULong -> Poke ()

peek :: Peek CULong

Store CLLong

Instance details

Methods

size :: Size CLLong

poke :: CLLong -> Poke ()

peek :: Peek CLLong

Store CULLong

Instance details

Methods

size :: Size CULLong

poke :: CULLong -> Poke ()

peek :: Peek CULLong

Store CBool

Instance details

Methods

size :: Size CBool

poke :: CBool -> Poke ()

peek :: Peek CBool

Store CFloat

Instance details

Methods

size :: Size CFloat

poke :: CFloat -> Poke ()

peek :: Peek CFloat

Store CDouble

Instance details

Methods

size :: Size CDouble

poke :: CDouble -> Poke ()

peek :: Peek CDouble

Store CPtrdiff

Instance details

Methods

size :: Size CPtrdiff

poke :: CPtrdiff -> Poke ()

peek :: Peek CPtrdiff

Store CSize

Instance details

Methods

size :: Size CSize

poke :: CSize -> Poke ()

peek :: Peek CSize

Store CWchar

Instance details

Methods

size :: Size CWchar

poke :: CWchar -> Poke ()

peek :: Peek CWchar

Store CSigAtomic

Instance details

Methods

size :: Size CSigAtomic

poke :: CSigAtomic -> Poke ()

peek :: Peek CSigAtomic

Store CClock

Instance details

Methods

size :: Size CClock

poke :: CClock -> Poke ()

peek :: Peek CClock

Store CTime

Instance details

Methods

size :: Size CTime

poke :: CTime -> Poke ()

peek :: Peek CTime

Store CUSeconds

Instance details

Methods

size :: Size CUSeconds

poke :: CUSeconds -> Poke ()

peek :: Peek CUSeconds

Store CSUSeconds

Instance details

Methods

size :: Size CSUSeconds

poke :: CSUSeconds -> Poke ()

peek :: Peek CSUSeconds

Store CIntPtr

Instance details

Methods

size :: Size CIntPtr

poke :: CIntPtr -> Poke ()

peek :: Peek CIntPtr

Store CUIntPtr

Instance details

Methods

size :: Size CUIntPtr

poke :: CUIntPtr -> Poke ()

peek :: Peek CUIntPtr

Store CIntMax

Instance details

Methods

size :: Size CIntMax

poke :: CIntMax -> Poke ()

peek :: Peek CIntMax

Store CUIntMax

Instance details

Methods

size :: Size CUIntMax

poke :: CUIntMax -> Poke ()

peek :: Peek CUIntMax

Store WordPtr

Instance details

Methods

size :: Size WordPtr

poke :: WordPtr -> Poke ()

peek :: Peek WordPtr

Store IntPtr

Instance details

Methods

size :: Size IntPtr

poke :: IntPtr -> Poke ()

peek :: Peek IntPtr

Store Fingerprint

Instance details

Methods

size :: Size Fingerprint

poke :: Fingerprint -> Poke ()

peek :: Peek Fingerprint

Store ShortByteString

Instance details

Methods

size :: Size ShortByteString

poke :: ShortByteString -> Poke ()

peek :: Peek ShortByteString

Store ByteString

Instance details

Methods

size :: Size ByteString

poke :: ByteString -> Poke ()

peek :: Peek ByteString

Store ByteString

Instance details

Methods

size :: Size ByteString

poke :: ByteString -> Poke ()

peek :: Peek ByteString

Store TimeSpec

Instance details

Methods

size :: Size TimeSpec

poke :: TimeSpec -> Poke ()

peek :: Peek TimeSpec

Store IntSet

Instance details

Methods

size :: Size IntSet

poke :: IntSet -> Poke ()

peek :: Peek IntSet

Store PortNumber

Instance details

Methods

size :: Size PortNumber

poke :: PortNumber -> Poke ()

peek :: Peek PortNumber

Store ModName

Instance details

Methods

size :: Size ModName

poke :: ModName -> Poke ()

peek :: Peek ModName

Store PkgName

Instance details

Methods

size :: Size PkgName

poke :: PkgName -> Poke ()

peek :: Peek PkgName

Store OccName

Instance details

Methods

size :: Size OccName

poke :: OccName -> Poke ()

peek :: Peek OccName

Store NameFlavour

Instance details

Methods

size :: Size NameFlavour

poke :: NameFlavour -> Poke ()

peek :: Peek NameFlavour

Store NameSpace

Instance details

Methods

size :: Size NameSpace

poke :: NameSpace -> Poke ()

peek :: Peek NameSpace

Store Info

Instance details

Methods

size :: Size Info

poke :: Info -> Poke ()

peek :: Peek Info

Store Fixity

Instance details

Methods

size :: Size Fixity

poke :: Fixity -> Poke ()

peek :: Peek Fixity

Store FixityDirection

Instance details

Methods

size :: Size FixityDirection

poke :: FixityDirection -> Poke ()

peek :: Peek FixityDirection

Store Lit

Instance details

Methods

size :: Size Lit

poke :: Lit -> Poke ()

peek :: Peek Lit

Store Bytes
Template Haskell Bytes are nearly identical to ByteString, but it can't depend on ByteString.

Instance details

Methods

size :: Size Bytes

poke :: Bytes -> Poke ()

peek :: Peek Bytes

Store Body

Instance details

Methods

size :: Size Body

poke :: Body -> Poke ()

peek :: Peek Body

Store Guard

Instance details

Methods

size :: Size Guard

poke :: Guard -> Poke ()

peek :: Peek Guard

Store Stmt

Instance details

Methods

size :: Size Stmt

poke :: Stmt -> Poke ()

peek :: Peek Stmt

Store Range

Instance details

Methods

size :: Size Range

poke :: Range -> Poke ()

peek :: Peek Range

Store DerivClause

Instance details

Methods

size :: Size DerivClause

poke :: DerivClause -> Poke ()

peek :: Peek DerivClause

Store DerivStrategy

Instance details

Methods

size :: Size DerivStrategy

poke :: DerivStrategy -> Poke ()

peek :: Peek DerivStrategy

Store TypeFamilyHead

Instance details

Methods

size :: Size TypeFamilyHead

poke :: TypeFamilyHead -> Poke ()

peek :: Peek TypeFamilyHead

Store TySynEqn

Instance details

Methods

size :: Size TySynEqn

poke :: TySynEqn -> Poke ()

peek :: Peek TySynEqn

Store Foreign

Instance details

Methods

size :: Size Foreign

poke :: Foreign -> Poke ()

peek :: Peek Foreign

Store Callconv

Instance details

Methods

size :: Size Callconv

poke :: Callconv -> Poke ()

peek :: Peek Callconv

Store Safety

Instance details

Methods

size :: Size Safety

poke :: Safety -> Poke ()

peek :: Peek Safety

Store Pragma

Instance details

Methods

size :: Size Pragma

poke :: Pragma -> Poke ()

peek :: Peek Pragma

Store Inline

Instance details

Methods

size :: Size Inline

poke :: Inline -> Poke ()

peek :: Peek Inline

Store RuleMatch

Instance details

Methods

size :: Size RuleMatch

poke :: RuleMatch -> Poke ()

peek :: Peek RuleMatch

Store Phases

Instance details

Methods

size :: Size Phases

poke :: Phases -> Poke ()

peek :: Peek Phases

Store RuleBndr

Instance details

Methods

size :: Size RuleBndr

poke :: RuleBndr -> Poke ()

peek :: Peek RuleBndr

Store AnnTarget

Instance details

Methods

size :: Size AnnTarget

poke :: AnnTarget -> Poke ()

peek :: Peek AnnTarget

Store SourceUnpackedness

Instance details

Methods

size :: Size SourceUnpackedness

poke :: SourceUnpackedness -> Poke ()

peek :: Peek SourceUnpackedness

Store SourceStrictness

Instance details

Methods

size :: Size SourceStrictness

poke :: SourceStrictness -> Poke ()

peek :: Peek SourceStrictness

Store Con

Instance details

Methods

size :: Size Con

poke :: Con -> Poke ()

peek :: Peek Con

Store Bang

Instance details

Methods

size :: Size Bang

poke :: Bang -> Poke ()

peek :: Peek Bang

Store PatSynDir

Instance details

Methods

size :: Size PatSynDir

poke :: PatSynDir -> Poke ()

peek :: Peek PatSynDir

Store PatSynArgs

Instance details

Methods

size :: Size PatSynArgs

poke :: PatSynArgs -> Poke ()

peek :: Peek PatSynArgs

Store TyVarBndr

Instance details

Methods

size :: Size TyVarBndr

poke :: TyVarBndr -> Poke ()

peek :: Peek TyVarBndr

Store FamilyResultSig

Instance details

Methods

size :: Size FamilyResultSig

poke :: FamilyResultSig -> Poke ()

peek :: Peek FamilyResultSig

Store TyLit

Instance details

Methods

size :: Size TyLit

poke :: TyLit -> Poke ()

peek :: Peek TyLit

Store Role

Instance details

Methods

size :: Size Role

poke :: Role -> Poke ()

peek :: Peek Role

Store Text

Instance details

Methods

size :: Size Text

poke :: Text -> Poke ()

peek :: Peek Text

Store FormatExtension

Instance details

Methods

size :: Size FormatExtension

poke :: FormatExtension -> Poke ()

peek :: Peek FormatExtension

Store ZonedTime

Instance details

Methods

size :: Size ZonedTime

poke :: ZonedTime -> Poke ()

peek :: Peek ZonedTime

Store TimeLocale

Instance details

Methods

size :: Size TimeLocale

poke :: TimeLocale -> Poke ()

peek :: Peek TimeLocale

Store LocalTime

Instance details

Methods

size :: Size LocalTime

poke :: LocalTime -> Poke ()

peek :: Peek LocalTime

Store TimeOfDay

Instance details

Methods

size :: Size TimeOfDay

poke :: TimeOfDay -> Poke ()

peek :: Peek TimeOfDay

Store TimeZone

Instance details

Methods

size :: Size TimeZone

poke :: TimeZone -> Poke ()

peek :: Peek TimeZone

Store CalendarDiffTime

Instance details

Methods

size :: Size CalendarDiffTime

poke :: CalendarDiffTime -> Poke ()

peek :: Peek CalendarDiffTime

Store UniversalTime

Instance details

Methods

size :: Size UniversalTime

poke :: UniversalTime -> Poke ()

peek :: Peek UniversalTime

Store UTCTime

Instance details

Methods

size :: Size UTCTime

poke :: UTCTime -> Poke ()

peek :: Peek UTCTime

Store SystemTime

Instance details

Methods

size :: Size SystemTime

poke :: SystemTime -> Poke ()

peek :: Peek SystemTime

Store NominalDiffTime

Instance details

Methods

size :: Size NominalDiffTime

poke :: NominalDiffTime -> Poke ()

peek :: Peek NominalDiffTime

Store AbsoluteTime

Instance details

Methods

size :: Size AbsoluteTime

poke :: AbsoluteTime -> Poke ()

peek :: Peek AbsoluteTime

Store DiffTime

Instance details

Methods

size :: Size DiffTime

poke :: DiffTime -> Poke ()

peek :: Peek DiffTime

Store Day

Instance details

Methods

size :: Size Day

poke :: Day -> Poke ()

peek :: Peek Day

Store CalendarDiffDays

Instance details

Methods

size :: Size CalendarDiffDays

poke :: CalendarDiffDays -> Poke ()

peek :: Peek CalendarDiffDays

Store a => Store [a]

Instance details

Methods

size :: Size [a]

poke :: [a] -> Poke ()

peek :: Peek [a]

Store a => Store (Maybe a)

Instance details

Methods

size :: Size (Maybe a)

poke :: Maybe a -> Poke ()

peek :: Peek (Maybe a)

Store a => Store (Ratio a)

Instance details

Methods

size :: Size (Ratio a)

poke :: Ratio a -> Poke ()

peek :: Peek (Ratio a)

Store (StablePtr a)

Instance details

Methods

size :: Size (StablePtr a)

poke :: StablePtr a -> Poke ()

peek :: Peek (StablePtr a)

Store (Ptr a)

Instance details

Methods

size :: Size (Ptr a)

poke :: Ptr a -> Poke ()

peek :: Peek (Ptr a)

Store (FunPtr a)

Instance details

Methods

size :: Size (FunPtr a)

poke :: FunPtr a -> Poke ()

peek :: Peek (FunPtr a)

Store a => Store (Complex a)

Instance details

Methods

size :: Size (Complex a)

poke :: Complex a -> Poke ()

peek :: Peek (Complex a)

Store a => Store (Identity a)

Instance details

Methods

size :: Size (Identity a)

poke :: Identity a -> Poke ()

peek :: Peek (Identity a)

Store a => Store (First a)

Instance details

Methods

size :: Size (First a)

poke :: First a -> Poke ()

peek :: Peek (First a)

Store a => Store (Last a)

Instance details

Methods

size :: Size (Last a)

poke :: Last a -> Poke ()

peek :: Peek (Last a)

Store a => Store (Dual a)

Instance details

Methods

size :: Size (Dual a)

poke :: Dual a -> Poke ()

peek :: Peek (Dual a)

Store a => Store (Sum a)

Instance details

Methods

size :: Size (Sum a)

poke :: Sum a -> Poke ()

peek :: Peek (Sum a)

Store a => Store (Product a)

Instance details

Methods

size :: Size (Product a)

poke :: Product a -> Poke ()

peek :: Peek (Product a)

Store a => Store (NonEmpty a)

Instance details

Methods

size :: Size (NonEmpty a)

poke :: NonEmpty a -> Poke ()

peek :: Peek (NonEmpty a)

Store a => Store (IntMap a)

Instance details

Methods

size :: Size (IntMap a)

poke :: IntMap a -> Poke ()

peek :: Peek (IntMap a)

Store a => Store (Seq a)

Instance details

Methods

size :: Size (Seq a)

poke :: Seq a -> Poke ()

peek :: Peek (Seq a)

(Store a, Ord a) => Store (Set a)

Instance details

Methods

size :: Size (Set a)

poke :: Set a -> Poke ()

peek :: Peek (Set a)

Prim a => Store (PrimStorable a)

Instance details

Methods

size :: Size (PrimStorable a)

poke :: PrimStorable a -> Poke ()

peek :: Peek (PrimStorable a)

(Eq a, Hashable a, Store a) => Store (HashSet a)

Instance details

Methods

size :: Size (HashSet a)

poke :: HashSet a -> Poke ()

peek :: Peek (HashSet a)

Store (Vector Bool)

Instance details

Methods

size :: Size (Vector Bool)

poke :: Vector Bool -> Poke ()

peek :: Peek (Vector Bool)

Store (Vector Char)

Instance details

Methods

size :: Size (Vector Char)

poke :: Vector Char -> Poke ()

peek :: Peek (Vector Char)

Store (Vector Double)

Instance details

Methods

size :: Size (Vector Double)

poke :: Vector Double -> Poke ()

peek :: Peek (Vector Double)

Store (Vector Float)

Instance details

Methods

size :: Size (Vector Float)

poke :: Vector Float -> Poke ()

peek :: Peek (Vector Float)

Store (Vector Int)

Instance details

Methods

size :: Size (Vector Int)

poke :: Vector Int -> Poke ()

peek :: Peek (Vector Int)

Store (Vector Int8)

Instance details

Methods

size :: Size (Vector Int8)

poke :: Vector Int8 -> Poke ()

peek :: Peek (Vector Int8)

Store (Vector Int16)

Instance details

Methods

size :: Size (Vector Int16)

poke :: Vector Int16 -> Poke ()

peek :: Peek (Vector Int16)

Store (Vector Int32)

Instance details

Methods

size :: Size (Vector Int32)

poke :: Vector Int32 -> Poke ()

peek :: Peek (Vector Int32)

Store (Vector Int64)

Instance details

Methods

size :: Size (Vector Int64)

poke :: Vector Int64 -> Poke ()

peek :: Peek (Vector Int64)

Store (Vector Word)

Instance details

Methods

size :: Size (Vector Word)

poke :: Vector Word -> Poke ()

peek :: Peek (Vector Word)

Store (Vector Word8)

Instance details

Methods

size :: Size (Vector Word8)

poke :: Vector Word8 -> Poke ()

peek :: Peek (Vector Word8)

Store (Vector Word16)

Instance details

Methods

size :: Size (Vector Word16)

poke :: Vector Word16 -> Poke ()

peek :: Peek (Vector Word16)

Store (Vector Word32)

Instance details

Methods

size :: Size (Vector Word32)

poke :: Vector Word32 -> Poke ()

peek :: Peek (Vector Word32)

Store (Vector Word64)

Instance details

Methods

size :: Size (Vector Word64)

poke :: Vector Word64 -> Poke ()

peek :: Peek (Vector Word64)

Store (Vector ())

Instance details

Methods

size :: Size (Vector ())

poke :: Vector () -> Poke ()

peek :: Peek (Vector ())

(Store (Vector a), Store (Vector b)) => Store (Vector (a, b))

Instance details

Methods

size :: Size (Vector (a, b))

poke :: Vector (a, b) -> Poke ()

peek :: Peek (Vector (a, b))

(Store (Vector a), Store (Vector b), Store (Vector c)) => Store (Vector (a, b, c))

Instance details

Methods

size :: Size (Vector (a, b, c))

poke :: Vector (a, b, c) -> Poke ()

peek :: Peek (Vector (a, b, c))

(Store (Vector a), Store (Vector b), Store (Vector c), Store (Vector d)) => Store (Vector (a, b, c, d))

Instance details

Methods

size :: Size (Vector (a, b, c, d))

poke :: Vector (a, b, c, d) -> Poke ()

peek :: Peek (Vector (a, b, c, d))

(Store (Vector a), Store (Vector b), Store (Vector c), Store (Vector d), Store (Vector e)) => Store (Vector (a, b, c, d, e))

Instance details

Methods

size :: Size (Vector (a, b, c, d, e))

poke :: Vector (a, b, c, d, e) -> Poke ()

peek :: Peek (Vector (a, b, c, d, e))

(Store (Vector a), Store (Vector b), Store (Vector c), Store (Vector d), Store (Vector e), Store (Vector f)) => Store (Vector (a, b, c, d, e, f))

Instance details

Methods

size :: Size (Vector (a, b, c, d, e, f))

poke :: Vector (a, b, c, d, e, f) -> Poke ()

peek :: Peek (Vector (a, b, c, d, e, f))

Store (Vector a) => Store (Vector (Complex a))

Instance details

Methods

size :: Size (Vector (Complex a))

poke :: Vector (Complex a) -> Poke ()

peek :: Peek (Vector (Complex a))

Store (Vector (f (g a))) => Store (Vector (Compose f g a))

Instance details

Methods

size :: Size (Vector (Compose f g a))

poke :: Vector (Compose f g a) -> Poke ()

peek :: Peek (Vector (Compose f g a))

Store (Vector a) => Store (Vector (Min a))

Instance details

Methods

size :: Size (Vector (Min a))

poke :: Vector (Min a) -> Poke ()

peek :: Peek (Vector (Min a))

Store (Vector a) => Store (Vector (Max a))

Instance details

Methods

size :: Size (Vector (Max a))

poke :: Vector (Max a) -> Poke ()

peek :: Peek (Vector (Max a))

(Store (Vector a), Store (Vector b)) => Store (Vector (Arg a b))

Instance details

Methods

size :: Size (Vector (Arg a b))

poke :: Vector (Arg a b) -> Poke ()

peek :: Peek (Vector (Arg a b))

Store (Vector a) => Store (Vector (First a))

Instance details

Methods

size :: Size (Vector (First a))

poke :: Vector (First a) -> Poke ()

peek :: Peek (Vector (First a))

Store (Vector a) => Store (Vector (Last a))

Instance details

Methods

size :: Size (Vector (Last a))

poke :: Vector (Last a) -> Poke ()

peek :: Peek (Vector (Last a))

Store (Vector a) => Store (Vector (WrappedMonoid a))

Instance details

Methods

size :: Size (Vector (WrappedMonoid a))

poke :: Vector (WrappedMonoid a) -> Poke ()

peek :: Peek (Vector (WrappedMonoid a))

Store (Vector a) => Store (Vector (Identity a))

Instance details

Methods

size :: Size (Vector (Identity a))

poke :: Vector (Identity a) -> Poke ()

peek :: Peek (Vector (Identity a))

Store (Vector a) => Store (Vector (Const a b))

Instance details

Methods

size :: Size (Vector (Const a b))

poke :: Vector (Const a b) -> Poke ()

peek :: Peek (Vector (Const a b))

Store (Vector a) => Store (Vector (Dual a))

Instance details

Methods

size :: Size (Vector (Dual a))

poke :: Vector (Dual a) -> Poke ()

peek :: Peek (Vector (Dual a))

Store (Vector All)

Instance details

Methods

size :: Size (Vector All)

poke :: Vector All -> Poke ()

peek :: Peek (Vector All)

Store (Vector Any)

Instance details

Methods

size :: Size (Vector Any)

poke :: Vector Any -> Poke ()

peek :: Peek (Vector Any)

Store (Vector a) => Store (Vector (Sum a))

Instance details

Methods

size :: Size (Vector (Sum a))

poke :: Vector (Sum a) -> Poke ()

peek :: Peek (Vector (Sum a))

Store (Vector a) => Store (Vector (Product a))

Instance details

Methods

size :: Size (Vector (Product a))

poke :: Vector (Product a) -> Poke ()

peek :: Peek (Vector (Product a))

Store (Vector (f a)) => Store (Vector (Alt f a))

Instance details

Methods

size :: Size (Vector (Alt f a))

poke :: Vector (Alt f a) -> Poke ()

peek :: Peek (Vector (Alt f a))

Store (Vector a) => Store (Vector (Down a))

Instance details

Methods

size :: Size (Vector (Down a))

poke :: Vector (Down a) -> Poke ()

peek :: Peek (Vector (Down a))

Storable a => Store (Vector a)

Instance details

Methods

size :: Size (Vector a)

poke :: Vector a -> Poke ()

peek :: Peek (Vector a)

Store (Vector Char)

Instance details

Methods

size :: Size (Vector Char)

poke :: Vector Char -> Poke ()

peek :: Peek (Vector Char)

Store (Vector Double)

Instance details

Methods

size :: Size (Vector Double)

poke :: Vector Double -> Poke ()

peek :: Peek (Vector Double)

Store (Vector Float)

Instance details

Methods

size :: Size (Vector Float)

poke :: Vector Float -> Poke ()

peek :: Peek (Vector Float)

Store (Vector Int)

Instance details

Methods

size :: Size (Vector Int)

poke :: Vector Int -> Poke ()

peek :: Peek (Vector Int)

Store (Vector Int8)

Instance details

Methods

size :: Size (Vector Int8)

poke :: Vector Int8 -> Poke ()

peek :: Peek (Vector Int8)

Store (Vector Int16)

Instance details

Methods

size :: Size (Vector Int16)

poke :: Vector Int16 -> Poke ()

peek :: Peek (Vector Int16)

Store (Vector Int32)

Instance details

Methods

size :: Size (Vector Int32)

poke :: Vector Int32 -> Poke ()

peek :: Peek (Vector Int32)

Store (Vector Int64)

Instance details

Methods

size :: Size (Vector Int64)

poke :: Vector Int64 -> Poke ()

peek :: Peek (Vector Int64)

Store (Vector (StablePtr a))

Instance details

Methods

size :: Size (Vector (StablePtr a))

poke :: Vector (StablePtr a) -> Poke ()

peek :: Peek (Vector (StablePtr a))

Store (Vector Word)

Instance details

Methods

size :: Size (Vector Word)

poke :: Vector Word -> Poke ()

peek :: Peek (Vector Word)

Store (Vector Word8)

Instance details

Methods

size :: Size (Vector Word8)

poke :: Vector Word8 -> Poke ()

peek :: Peek (Vector Word8)

Store (Vector Word16)

Instance details

Methods

size :: Size (Vector Word16)

poke :: Vector Word16 -> Poke ()

peek :: Peek (Vector Word16)

Store (Vector Word32)

Instance details

Methods

size :: Size (Vector Word32)

poke :: Vector Word32 -> Poke ()

peek :: Peek (Vector Word32)

Store (Vector Word64)

Instance details

Methods

size :: Size (Vector Word64)

poke :: Vector Word64 -> Poke ()

peek :: Peek (Vector Word64)

Store (Vector (Ptr a))

Instance details

Methods

size :: Size (Vector (Ptr a))

poke :: Vector (Ptr a) -> Poke ()

peek :: Peek (Vector (Ptr a))

Store (Vector (FunPtr a))

Instance details

Methods

size :: Size (Vector (FunPtr a))

poke :: Vector (FunPtr a) -> Poke ()

peek :: Peek (Vector (FunPtr a))

Prim a => Store (Vector (Min a))

Instance details

Methods

size :: Size (Vector (Min a))

poke :: Vector (Min a) -> Poke ()

peek :: Peek (Vector (Min a))

Prim a => Store (Vector (Max a))

Instance details

Methods

size :: Size (Vector (Max a))

poke :: Vector (Max a) -> Poke ()

peek :: Peek (Vector (Max a))

Prim a => Store (Vector (First a))

Instance details

Methods

size :: Size (Vector (First a))

poke :: Vector (First a) -> Poke ()

peek :: Peek (Vector (First a))

Prim a => Store (Vector (Last a))

Instance details

Methods

size :: Size (Vector (Last a))

poke :: Vector (Last a) -> Poke ()

peek :: Peek (Vector (Last a))

Prim a => Store (Vector (Identity a))

Instance details

Methods

size :: Size (Vector (Identity a))

poke :: Vector (Identity a) -> Poke ()

peek :: Peek (Vector (Identity a))

Store (Vector CDev)

Instance details

Methods

size :: Size (Vector CDev)

poke :: Vector CDev -> Poke ()

peek :: Peek (Vector CDev)

Store (Vector CIno)

Instance details

Methods

size :: Size (Vector CIno)

poke :: Vector CIno -> Poke ()

peek :: Peek (Vector CIno)

Store (Vector CMode)

Instance details

Methods

size :: Size (Vector CMode)

poke :: Vector CMode -> Poke ()

peek :: Peek (Vector CMode)

Store (Vector COff)

Instance details

Methods

size :: Size (Vector COff)

poke :: Vector COff -> Poke ()

peek :: Peek (Vector COff)

Store (Vector CPid)

Instance details

Methods

size :: Size (Vector CPid)

poke :: Vector CPid -> Poke ()

peek :: Peek (Vector CPid)

Store (Vector CSsize)

Instance details

Methods

size :: Size (Vector CSsize)

poke :: Vector CSsize -> Poke ()

peek :: Peek (Vector CSsize)

Store (Vector CGid)

Instance details

Methods

size :: Size (Vector CGid)

poke :: Vector CGid -> Poke ()

peek :: Peek (Vector CGid)

Store (Vector CNlink)

Instance details

Methods

size :: Size (Vector CNlink)

poke :: Vector CNlink -> Poke ()

peek :: Peek (Vector CNlink)

Store (Vector CUid)

Instance details

Methods

size :: Size (Vector CUid)

poke :: Vector CUid -> Poke ()

peek :: Peek (Vector CUid)

Store (Vector CCc)

Instance details

Methods

size :: Size (Vector CCc)

poke :: Vector CCc -> Poke ()

peek :: Peek (Vector CCc)

Store (Vector CSpeed)

Instance details

Methods

size :: Size (Vector CSpeed)

poke :: Vector CSpeed -> Poke ()

peek :: Peek (Vector CSpeed)

Store (Vector CTcflag)

Instance details

Methods

size :: Size (Vector CTcflag)

poke :: Vector CTcflag -> Poke ()

peek :: Peek (Vector CTcflag)

Store (Vector CRLim)

Instance details

Methods

size :: Size (Vector CRLim)

poke :: Vector CRLim -> Poke ()

peek :: Peek (Vector CRLim)

Store (Vector CBlkSize)

Instance details

Methods

size :: Size (Vector CBlkSize)

poke :: Vector CBlkSize -> Poke ()

peek :: Peek (Vector CBlkSize)

Store (Vector CBlkCnt)

Instance details

Methods

size :: Size (Vector CBlkCnt)

poke :: Vector CBlkCnt -> Poke ()

peek :: Peek (Vector CBlkCnt)

Store (Vector CClockId)

Instance details

Methods

size :: Size (Vector CClockId)

poke :: Vector CClockId -> Poke ()

peek :: Peek (Vector CClockId)

Store (Vector CFsBlkCnt)

Instance details

Methods

size :: Size (Vector CFsBlkCnt)

poke :: Vector CFsBlkCnt -> Poke ()

peek :: Peek (Vector CFsBlkCnt)

Store (Vector CFsFilCnt)

Instance details

Methods

size :: Size (Vector CFsFilCnt)

poke :: Vector CFsFilCnt -> Poke ()

peek :: Peek (Vector CFsFilCnt)

Store (Vector CId)

Instance details

Methods

size :: Size (Vector CId)

poke :: Vector CId -> Poke ()

peek :: Peek (Vector CId)

Store (Vector CKey)

Instance details

Methods

size :: Size (Vector CKey)

poke :: Vector CKey -> Poke ()

peek :: Peek (Vector CKey)

Store (Vector CTimer)

Instance details

Methods

size :: Size (Vector CTimer)

poke :: Vector CTimer -> Poke ()

peek :: Peek (Vector CTimer)

Store (Vector Fd)

Instance details

Methods

size :: Size (Vector Fd)

poke :: Vector Fd -> Poke ()

peek :: Peek (Vector Fd)

Prim a => Store (Vector (Const a b))

Instance details

Methods

size :: Size (Vector (Const a b))

poke :: Vector (Const a b) -> Poke ()

peek :: Peek (Vector (Const a b))

Prim a => Store (Vector (Dual a))

Instance details

Methods

size :: Size (Vector (Dual a))

poke :: Vector (Dual a) -> Poke ()

peek :: Peek (Vector (Dual a))

Prim a => Store (Vector (Sum a))

Instance details

Methods

size :: Size (Vector (Sum a))

poke :: Vector (Sum a) -> Poke ()

peek :: Peek (Vector (Sum a))

Prim a => Store (Vector (Product a))

Instance details

Methods

size :: Size (Vector (Product a))

poke :: Vector (Product a) -> Poke ()

peek :: Peek (Vector (Product a))

Prim a => Store (Vector (Down a))

Instance details

Methods

size :: Size (Vector (Down a))

poke :: Vector (Down a) -> Poke ()

peek :: Peek (Vector (Down a))

Store (Vector CChar)

Instance details

Methods

size :: Size (Vector CChar)

poke :: Vector CChar -> Poke ()

peek :: Peek (Vector CChar)

Store (Vector CSChar)

Instance details

Methods

size :: Size (Vector CSChar)

poke :: Vector CSChar -> Poke ()

peek :: Peek (Vector CSChar)

Store (Vector CUChar)

Instance details

Methods

size :: Size (Vector CUChar)

poke :: Vector CUChar -> Poke ()

peek :: Peek (Vector CUChar)

Store (Vector CShort)

Instance details

Methods

size :: Size (Vector CShort)

poke :: Vector CShort -> Poke ()

peek :: Peek (Vector CShort)

Store (Vector CUShort)

Instance details

Methods

size :: Size (Vector CUShort)

poke :: Vector CUShort -> Poke ()

peek :: Peek (Vector CUShort)

Store (Vector CInt)

Instance details

Methods

size :: Size (Vector CInt)

poke :: Vector CInt -> Poke ()

peek :: Peek (Vector CInt)

Store (Vector CUInt)

Instance details

Methods

size :: Size (Vector CUInt)

poke :: Vector CUInt -> Poke ()

peek :: Peek (Vector CUInt)

Store (Vector CLong)

Instance details

Methods

size :: Size (Vector CLong)

poke :: Vector CLong -> Poke ()

peek :: Peek (Vector CLong)

Store (Vector CULong)

Instance details

Methods

size :: Size (Vector CULong)

poke :: Vector CULong -> Poke ()

peek :: Peek (Vector CULong)

Store (Vector CLLong)

Instance details

Methods

size :: Size (Vector CLLong)

poke :: Vector CLLong -> Poke ()

peek :: Peek (Vector CLLong)

Store (Vector CULLong)

Instance details

Methods

size :: Size (Vector CULLong)

poke :: Vector CULLong -> Poke ()

peek :: Peek (Vector CULLong)

Store (Vector CBool)

Instance details

Methods

size :: Size (Vector CBool)

poke :: Vector CBool -> Poke ()

peek :: Peek (Vector CBool)

Store (Vector CFloat)

Instance details

Methods

size :: Size (Vector CFloat)

poke :: Vector CFloat -> Poke ()

peek :: Peek (Vector CFloat)

Store (Vector CDouble)

Instance details

Methods

size :: Size (Vector CDouble)

poke :: Vector CDouble -> Poke ()

peek :: Peek (Vector CDouble)

Store (Vector CPtrdiff)

Instance details

Methods

size :: Size (Vector CPtrdiff)

poke :: Vector CPtrdiff -> Poke ()

peek :: Peek (Vector CPtrdiff)

Store (Vector CSize)

Instance details

Methods

size :: Size (Vector CSize)

poke :: Vector CSize -> Poke ()

peek :: Peek (Vector CSize)

Store (Vector CWchar)

Instance details

Methods

size :: Size (Vector CWchar)

poke :: Vector CWchar -> Poke ()

peek :: Peek (Vector CWchar)

Store (Vector CSigAtomic)

Instance details

Methods

size :: Size (Vector CSigAtomic)

poke :: Vector CSigAtomic -> Poke ()

peek :: Peek (Vector CSigAtomic)

Store (Vector CClock)

Instance details

Methods

size :: Size (Vector CClock)

poke :: Vector CClock -> Poke ()

peek :: Peek (Vector CClock)

Store (Vector CTime)

Instance details

Methods

size :: Size (Vector CTime)

poke :: Vector CTime -> Poke ()

peek :: Peek (Vector CTime)

Store (Vector CUSeconds)

Instance details

Methods

size :: Size (Vector CUSeconds)

poke :: Vector CUSeconds -> Poke ()

peek :: Peek (Vector CUSeconds)

Store (Vector CSUSeconds)

Instance details

Methods

size :: Size (Vector CSUSeconds)

poke :: Vector CSUSeconds -> Poke ()

peek :: Peek (Vector CSUSeconds)

Store (Vector CIntPtr)

Instance details

Methods

size :: Size (Vector CIntPtr)

poke :: Vector CIntPtr -> Poke ()

peek :: Peek (Vector CIntPtr)

Store (Vector CUIntPtr)

Instance details

Methods

size :: Size (Vector CUIntPtr)

poke :: Vector CUIntPtr -> Poke ()

peek :: Peek (Vector CUIntPtr)

Store (Vector CIntMax)

Instance details

Methods

size :: Size (Vector CIntMax)

poke :: Vector CIntMax -> Poke ()

peek :: Peek (Vector CIntMax)

Store (Vector CUIntMax)

Instance details

Methods

size :: Size (Vector CUIntMax)

poke :: Vector CUIntMax -> Poke ()

peek :: Peek (Vector CUIntMax)

Store (Vector WordPtr)

Instance details

Methods

size :: Size (Vector WordPtr)

poke :: Vector WordPtr -> Poke ()

peek :: Peek (Vector WordPtr)

Store (Vector IntPtr)

Instance details

Methods

size :: Size (Vector IntPtr)

poke :: Vector IntPtr -> Poke ()

peek :: Peek (Vector IntPtr)

Store a => Store (Vector a)

Instance details

Methods

size :: Size (Vector a)

poke :: Vector a -> Poke ()

peek :: Peek (Vector a)

(Store a, Store b) => Store (Either a b)

Instance details

Methods

size :: Size (Either a b)

poke :: Either a b -> Poke ()

peek :: Peek (Either a b)

(Store a, Store b) => Store (a, b)

Instance details

Methods

size :: Size (a, b)

poke :: (a, b) -> Poke ()

peek :: Peek (a, b)

(Ix i, IArray UArray e, Store i, Store e) => Store (UArray i e)

Instance details

Methods

size :: Size (UArray i e)

poke :: UArray i e -> Poke ()

peek :: Peek (UArray i e)

(Ix i, Store i, Store e) => Store (Array i e)

Instance details

Methods

size :: Size (Array i e)

poke :: Array i e -> Poke ()

peek :: Peek (Array i e)

Store (Fixed a)

Instance details

Methods

size :: Size (Fixed a)

poke :: Fixed a -> Poke ()

peek :: Peek (Fixed a)

(Ord k, Store k, Store a) => Store (Map k a)

Instance details

Methods

size :: Size (Map k a)

poke :: Map k a -> Poke ()

peek :: Peek (Map k a)

(Eq k, Hashable k, Store k, Store a) => Store (HashMap k a)

Instance details

Methods

size :: Size (HashMap k a)

poke :: HashMap k a -> Poke ()

peek :: Peek (HashMap k a)

(Store a, Store b, Store c) => Store (a, b, c)

Instance details

Methods

size :: Size (a, b, c)

poke :: (a, b, c) -> Poke ()

peek :: Peek (a, b, c)

Store a => Store (Const a b)

Instance details

Methods

size :: Size (Const a b)

poke :: Const a b -> Poke ()

peek :: Peek (Const a b)

(Store a, Store b, Store c, Store d) => Store (a, b, c, d)

Instance details

Methods

size :: Size (a, b, c, d)

poke :: (a, b, c, d) -> Poke ()

peek :: Peek (a, b, c, d)

(Store a, Store b, Store c, Store d, Store e) => Store (a, b, c, d, e)

Instance details

Methods

size :: Size (a, b, c, d, e)

poke :: (a, b, c, d, e) -> Poke ()

peek :: Peek (a, b, c, d, e)

(Store a, Store b, Store c, Store d, Store e, Store f) => Store (a, b, c, d, e, f)

Instance details

Methods

size :: Size (a, b, c, d, e, f)

poke :: (a, b, c, d, e, f) -> Poke ()

peek :: Peek (a, b, c, d, e, f)

(Store a, Store b, Store c, Store d, Store e, Store f, Store g) => Store (a, b, c, d, e, f, g)

Instance details

Methods

size :: Size (a, b, c, d, e, f, g)

poke :: (a, b, c, d, e, f, g) -> Poke ()

peek :: Peek (a, b, c, d, e, f, g)