Data.Store.TypeHash.Internal

Documentation

newtype Tagged a

Constructors

Tagged

Fields

unTagged :: a

newtype TypeHash

Constructors

TypeHash

Fields

unTypeHash :: StaticSize 20 ByteString

reifyManyTyDecls :: ((Name, Info) -> Q (Bool, [Name])) -> [Name] -> Q [(Name, Info)]

typeHashForNames :: [Name] -> Q Exp

At compiletime, this yields a hash of the specified datatypes. Not only does this cover the datatypes themselves, but also all transitive dependencies.

The resulting expression is a literal of type TypeHash.

hashOfType :: Type -> Q Exp

At compiletime, this yields a cryptographic hash of the specified Type, including the definition of things it references (transitively).

The resulting expression is a literal of type TypeHash.

getTypeInfosRecursively :: [Name] -> Q [(Name, Info)]

getConNames :: Data a => a -> [Name]

getVarNames :: Data a => a -> [Name]

class HasTypeHash a where

Methods

typeHash :: Proxy a -> TypeHash

mkHasTypeHash :: Type -> Q [Dec]

Deprecated: Use of Data.Store.TypeHash isn't recommended, as the hashes are too unstable for most uses. Please instead consider using Data.Store.Version. See https://github.com/fpco/store/issues/53

mkManyHasTypeHash :: [Q Type] -> Q [Dec]

Deprecated: Use of Data.Store.TypeHash isn't recommended, as the hashes are too unstable for most uses. Please instead consider using Data.Store.Version. See https://github.com/fpco/store/issues/53

combineTypeHashes :: [TypeHash] -> TypeHash